separated container logic from presentation in employee search list. Employee components now fully separated
